server/requirements.txt
aiodns==3.0.0
alembic==1.11.1
annotated-types==0.5.0
asgiref==3.7.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
email-normalize==2.0.0
exceptiongroup==1.1.2
Flask==2.3.2
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.0.5
google-auth==2.22.0
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycares==4.3.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
PyJWT==2.8.0
pytest==7.4.0
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.19
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.6
zipp==3.16.2
